+++ sovellus/src/services/counter_service.py
@@ -1,3 +1,5 @@
+"""Sovelluslogiikasta vastaavat luokat."""
+
 from entities.counter import Counter
 from entities.user import User
 
@@ -12,34 +14,42 @@ from repositories.user_repository import (
 class InvalidCredentialsError(Exception):
     pass
 
-
 class UsernameExistsError(Exception):
     pass
 
 class CounterService:
+    """Luokka, joka vastaa sovelluslogiikasta aurinkopaneelilaskurissa."""
+
     def __init__(
         self,
         counter_repository=default_counter_repository,
         user_repository=default_user_repository
     ):
+        """Luokan konstruktori. Luo uuden sovelluslogiikasta vastaavan palvelun.
+
+        Args:
+            counter_repository:
+            Vapaaehtoinen, oletusarvoisesti CounterRepository-olio.
+            Olio, joka vastaa laskelmien tallennuksesta ja hakemisesta.
+            user_repository:
+            Vapaaehtoinen, oletusarvoisesti UserRepository-olio.
+            Olio, joka vastaa käyttäjiin liittyvistä tietokantaoperaatioista.
+        """
         self._user = None
         self._counter_repository = counter_repository
         self._user_repository = user_repository
 
     def calculate_net_consumption(self, e_consumption, e_generation):
+        """Laskee nettokulutuksen.
+
+        Args:
+            e_consumption: Sähkön vuosittainen kulutus (kWh).
+            e_generation: Oman aurinkopaneelin vuosittainen sähköntuotto (kWh).
+
+        Returns:
+            Nettokulutus (sähkönkulutus - sähköntuotto)
+        """
         return e_consumption - e_generation
-
-    def calculate_savings(self, e_generation, e_purchase_price):
-        return e_generation * e_purchase_price
-
-    def calculate_payback_time(self, net_cost, yearly_savings):
-        return net_cost / yearly_savings
-
-    def calculate(self, acquisition_cost, tax_deduction, e_generation, e_purchase_price):
-        net_cost = acquisition_cost - tax_deduction
-        yearly_savings = self.calculate_savings(e_generation, e_purchase_price)
-        payback_time = self.calculate_payback_time(net_cost, yearly_savings)
-        return yearly_savings, payback_time
 
     def save_calculation(self,
                          user,
@@ -48,28 +58,49 @@ class CounterService:
                          e_consumption,
                          e_generation,
                          e_purchase_price):
+        """Tallentaa käyttäjän aurinkopaneelilaskelman tietokantaan.
+
+        Args:
+            user: Käyttäjä, jolle laskelma kuuluu.
+            acquisition_cost: Aurinkopaneelien hankintahinta (eur).
+            tax_deduction: Kotitalousvähennys hankintahinnasta (eur).
+            e_consumption: Vuosittainen sähkönkulutus (kWh).
+            e_generation: Vuosittainen sähköntuotto (kWh).
+            e_purchase_price: Sähkön ostohinta (eur/kWh).
+
+        Returns:
+            Counter: Tallennettu laskelma Counter-oliona.
+        """
         calculation = Counter(
             acquisition_cost=acquisition_cost,
             tax_deduction=tax_deduction,
             e_consumption=e_consumption,
             e_generation=e_generation,
             e_purchase_price=e_purchase_price,
-            user_id=user.id
+            user_id=user.user_id
         )
-
-        yearly_savings, payback_time = self.calculate(
-            acquisition_cost,
-            tax_deduction,
-            e_generation,
-            e_purchase_price
-        )
-
-        calculation.yearly_savings = yearly_savings
-        calculation.payback_time = payback_time
 
         return self._counter_repository.save_counter(calculation)
 
     def create_user(self, username, password, login=True):
+        """Luo uuden käyttäjän ja kirjaa tarvittaessa sisään.
+
+        Args:
+            username:
+            Käyttäjän käyttäjätunnus (merkkijonoarvo)
+            password:
+            Käyttäjän salasana (merkkijonoarvo)
+            login:
+            Vapaaehtoinen, oletusarvo True.
+            Kertoo kirjataanko käyttäjä sisään onnistuneen luonnin jälkeen (boolean).
+
+        Raises:
+            UsernameExistsError:
+            Virhe, joka heitetään, jos käyttäjätunnus on jo olemassa.
+
+        Returns:
+            Luotu käyttäjä User-olion muodossa.
+        """
 
         existing_user = self._user_repository.find_by_username(username)
 
@@ -84,9 +115,6 @@ class CounterService:
         return user
 
     def login(self, username, password):
-        #poista
-        print(f"Yritetään kirjautua sisään käyttäjällä: {username}")
-        #poista
 
         user = self._user_repository.find_by_username(username)
 
@@ -98,20 +126,54 @@ class CounterService:
             print("Salasana väärin.")
             raise InvalidCredentialsError("Virheellinen käyttäjätunnus tai salasana")
 
-        print(f"Kirjautunut käyttäjä: {user.username}")
         self._user = user
 
         return user
 
     def get_user(self):
+        """Palauttaa tällä hetkellä sisäänkirjatun käyttäjän.
+
+        Returns:
+            User: Tällä hetkellä sisäänkirjatun käyttäjän User-olio.
+                  Jos käyttäjää ei ole kirjautuneena, palautetaan None.
+        """
 
         return self._user
 
     def get_users(self):
+        """Palauttaa kaikki käyttäjät tietokannasta.
+
+        Returns:
+            User-oliota sisältä lista kaikista tietokannassa olevista käyttäjistä.
+        """
 
         return self._user_repository.find_all()
 
+    def get_previous_calculations(self, user):
+        """Palauttaa kaikki aiemmat laskelmat annetulle käyttäjälle.
+
+        Args:
+            user: Käyttäjä, jonka aiemmat laskelmat haetaan.
+
+        Returns:
+            Lista käyttäjän aiemmista laskelmista Counter-olioina.
+        """
+
+        return self._counter_repository.find_by_user_id(user.user_id)
+
+    def delete_counter(self, counter_id):
+        """Poistaa laskelman tietokannasta.
+
+        Args:
+            counter_id: Poistettavan laskelman id.
+
+        Returns:
+           Palauttaa True, jos laskelman poistaminen onnistui, muuten False.
+        """
+        return self._counter_repository.delete(counter_id)
+
     def logout(self):
+        """Kirjaa nykyisen käyttäjän ulos sovelluksesta."""
 
         self._user = None