+++ sovellus/dokumentaatio/changelog.md
@@ -10,11 +10,18 @@
 - Tietokannan perustaminen
 - Käyttäjä pystyy itse syöttämään laskuriin vaadittavat tiedot ja näkee näiden tietojen avulla lasketut säästöt ja kuoletusajan
 - Käyttäjä näkee listan kaikista tietokantaan tallennetuista laskelmista (ei vielä käyttäjärajausta)
-- Käyttäjä pystyy luomaan uuden tunnuksen ja salasanan. Käyttäjätunnus ja salasana tallennetaan tietokantaan. 
+- Käyttäjä pystyy luomaan uuden tunnuksen ja salasanan. Käyttäjätunnus ja salasana tallennetaan tietokantaan.
 
 ## Viikko 5
 
 - UI CreateUserView
 - UI LoginView
-- UI CounterView 
+- UI CounterView
 - Käyttäjä voi kirjautua sovellukseen sisälle ja ulos
+
+## Viikko 6
+
+- Laskuriin lisätty kustannus-kohta, joka päivittyy automaattisesti, kun hankintahinta ja verovähennys syötetään
+- Käyttäjä näkee aiemmat laskelmansa
+- Poistomahdollisuus aiempiin laskelmiin
+